Honour job dependencies
The order of jobs now is based on their dependencies (and the order of
their id where no dependency dictates an order).
If a job depends on a job that finished with failure, then that job
fails as well.

If a job depends on a job that is unknown, then that, too, means the job
fails. This should be prevented when saving the project already, by
validating the job descriptions and their dependencies.