Refactor copy scheduling
schedule_build() resulted in a build save before BuildSource existed, so
the row in the client wouldn't give the reason with the clickable link
to the source build until a worker scheduled it and therefore caused
another save signal.

This makes sure the BuildSource exists when schedule_build is called.

The client still gets far too many events due to the save signal. It's
probably a good idea to have a separate signal to publish to the
websocket to reduce redundant traffic.