Try to become functional again after web socket disconnect
Since there is no backend support for restoring the whole list or make
sure the build detail page state matches the backend, it is easiest to
reload the page once the websocket connected again. This, of course,
will require another connection, but it prevents the site from trying
to reload while the server isn't reachable for whatever reason.